fix: vulnerability allowing login in any account with any existing deviceId

+++ src/Whisper.Client/src/api/agent.ts
@@ -32,7 +32,7 @@ agent.interceptors.request.use(async (config: InternalAxiosRequestConfig) => {
 agent.interceptors.response.use(
     (response) => response,
     async (error) => {
-        if (error.response && error.response.status === 401) {
+        if (error.response && error.response.status === 401 && window.location.pathname !== "/auth/login") {
             await clearAuthData();
             localStorage.clear();
             window.location.replace("/auth/login");
